requirements.txt
# Django and REST Framework
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=23.5,<24.0

# Database
psycopg2-binary>=2.9,<3.0
dj-database-url>=2.1,<3.0

# Authentication
djangorestframework-simplejwt>=5.3,<6.0

# Storage
django-storages>=1.14,<2.0
boto3>=1.34,<2.0

# Production server
gunicorn>=21.2,<22.0
whitenoise>=6.6,<7.0

# Utilities
python-decouple>=3.8,<4.0
Pillow>=10.2,<11.0

# Celery (optional, for background tasks)
celery>=5.3,<6.0
redis>=5.0,<6.0

# Development
django-debug-toolbar>=4.2,<5.0

